+++ packages/svelte/src/compiler/phases/3-transform/client/visitors/javascript-runes.js
@@ -195,8 +195,7 @@ export const javascript_visitors_runes = {
 			if (rune === '$props') {
 				assert.equal(declarator.id.type, 'ObjectPattern');
 
-				/** @type {string[]} */
-				const seen = [];
+				const seen = state.analysis.exports.map(({ name, alias }) => alias ?? name);
 
 				for (const property of declarator.id.properties) {
 					if (property.type === 'Property') {
@@ -380,6 +379,10 @@ export const javascript_visitors_runes = {
 	},
 	CallExpression(node, context) {
 		const rune = get_rune(node, context.state.scope);
+
+		if (rune === '$host') {
+			return b.id('$$props.$$host');
+		}
 
 		if (rune === '$effect.active') {
 			return b.call('$.effect_active');