+++ packages/svelte/tests/runtime-legacy/samples/context-in-template/main.svelte
@@ -1,0 +1,11 @@
+<script>
+	import { getContext, setContext } from 'svelte';
+
+	setContext('val', 'hello world')
+
+	const get_val = () => {
+		return getContext('val');
+	}
+</script>
+
+{get_val()}